pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watchtower"
version = "0.1.0"
description = "Hourly market data watcher with SQLite storage and Streamlit dashboard."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Watchtower", email = "you@example.com" },
]
dependencies = [
  "yfinance>=0.2.50",
  "pandas>=2.0.0",
  "streamlit>=1.40.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "jobs*", "alerts*", "scripts*"]

